Fixed Cube Resizing / Rewrote Line Drawing
Updated the rotate function to avoid the cube shrinking as it spins.
Rewrote the line drawing algorithm. It is slightly different than before, but seems to have the same functionality. I will continue to optimize this, but as of right now it is in a working state and I'm happy with it.